Checkpointing the code because it's working, so before I start messing with it and break it again I should commit.

+++ src/drivers/waveshare-api/include/waveshare_api.h
@@ -38,7 +38,7 @@ waveshareapi_error waveshareapi_moduleinit();
 
 void waveshareapi_moduleexit();
 
-void waveshareapi_render_region(imagebuffer_t *buffer, uint16_t x_start, uint16_t y_start, uint16_t x_end, uint16_t y_end, uint8_t rotation);
+void waveshareapi_render_region(imagebuffer_t *buffer, uint16_t x_start, uint16_t y_start, uint16_t x_end, uint16_t y_end, uint16_t rotation);
 
 void waveshareapi_clear(uint8_t colour);
